Archive/training/requirements.txt
# Training module specific requirements
# Install with: pip install -r training/requirements.txt

# Core MLX requirements
mlx>=0.5.0
mlx-lm>=0.10.0

# Tokenizers and model handling
transformers>=4.36.0
tokenizers>=0.15.0
safetensors>=0.4.0

# Data handling
numpy>=1.24.0
pyyaml>=6.0.1
tqdm>=4.66.0

# Monitoring and visualization
tensorboard>=2.15.0
wandb>=0.16.0
matplotlib>=3.7.0

# System monitoring
psutil>=5.9.0

# Optional but recommended
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0